+++ packages/sdk-browser/src/sdk-client.ts
@@ -531,7 +531,6 @@ export class TurnkeyPasskeyClient extends TurnkeyBrowserClient {
   rpId: string;
 
   constructor(config: TurnkeySDKClientConfig) {
-    console.log("config", config);
     super(config, AuthClient.Passkey);
     this.rpId = (this.stamper as WebauthnStamper)!.rpId;
   }