+++ packages/sdk-browser/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @turnkey/sdk-browser
 
+## 1.13.0
+
+### Minor Changes
+
+- 57f9cb0: Update endpoints - surface GetWalletAccount
+
+### Patch Changes
+
+- 69d2571: Upgrade elliptic
+- Updated dependencies [57f9cb0]
+  - @turnkey/http@2.19.0
+  - @turnkey/crypto@2.3.1
+  - @turnkey/wallet-stamper@1.0.3
+
+## 1.12.1
+
+### Patch Changes
+
+- 755833b: refactor stamper out of config object and move it directly onto the client to match @turnkey/http
+- Updated dependencies [2bc0046]
+  - @turnkey/crypto@2.3.1
+  - @turnkey/wallet-stamper@1.0.3
+
+## 1.12.0
+
+### Minor Changes
+
+- 6695af2: Update per mono release v2025.1.11
+
+### Patch Changes
+
+- Updated dependencies [6695af2]
+  - @turnkey/http@2.18.0
+  - @turnkey/crypto@2.3.0
+  - @turnkey/wallet-stamper@1.0.2
+
 ## 1.11.2
 
 ### Patch Changes